+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.10.0] - 2023-xx-xx
 
 ### Added
+- GatewayMgmt: support for Dim=22 DateTime and parseDateTime()/toOWNDateTime() static methods and related tests
 - Energy: support for partial totalizer for current month and current day (WHO=18)
 - WereThermo: support for 4-zone central unit (where=#0#1) and related test
 
@@ -11,6 +12,7 @@
 - updated dependencies: slf4j-api, junit-jupiter-engine, maven-compiler-plugin, maven-javadoc-plugin, maven-jar-plugin, maven-bundle-plugin, maven-surefire-plugin, spotless-maven-plugin, nexus-staging-maven-plugin
 - DEPRECATED WhereThermo.isStandAlone()
 - updated BUSGateway to discover and notify thermo CU first and then zones
+- WhereThermo: deprected WhereThermo.isStandAlone()
 - escaped OWN frames in log messages